fix broken tests, round 1
