+++ src/components/Nav/Nav.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { Link } from "react-router-dom";
 import './Nav.scss';
 
-function Nav({ userInfo }) {
+function Nav(props) {
 
   return <nav>
     <section>
@@ -14,8 +14,8 @@ function Nav({ userInfo }) {
     </Link>
     </section>
     <section>
-      <p>Welcome, {userInfo.userName || "friend!"}</p>
-      <p>Your Account Type: {userInfo.accountType || "Other"}</p>
+      <p>Welcome, {props.userName || "friend!"}</p>
+      <p>Your Account Type: {props.accountType || "Other"}</p>
     </section>
     <Link to="/">
       <button>Sign Out</button>
